requirements.txt
discord.py
asyncpg
pillow
python-dotenv
